Refactor query execution modules
